telefono and user

+++ models/telefono.js
@@ -4,7 +4,7 @@ const user = require('./user');
 
 const telefono = db.define('Telefono', {
     telefono: {
-        type: sequelize.INTEGER,
+        type: sequelize.DOUBLE,
         allowNull: false,
         validate: {
             notEmpty: true,
